+++ R/sim_game_abr.R
@@ -34,51 +34,69 @@ sim_game_abr <- function(game,
 
   payoff1 <- payoff2 <- NULL
 
-  n2 <- n_periods * 2
+  np2 <- n_periods * 2
 
-  play1 <- rep(NA, n2)
-  play2 <- rep(NA, n2)
+  play1 <- rep(NA, np2)
+  play2 <- rep(NA, np2)
 
   if (game$type == "matrix") {
 
     s1 <- game$strategy$s1
     s2 <- game$strategy$s2
 
-    # for the first round
-    if (is.null(init1)) play1[1] <- sample(s1, size = 1)
-    else play1[1] <- init1
-    if (is.null(init2)) play2[1] <- sample(s2, size = 1)
-    else play2[1] <- init2
+    n1 <- length(s1)
+    n2 <- length(s2)
 
-    for (i in 2:n2) {
-      if (i %% 2 != 0) {
+    pi1 <- game$mat$matrix1
+    pi2 <- game$mat$matrix2
+
+    # for the first round
+    if (is.null(init1)) {
+      play1[1] <- sample(1:n1, size = 1)
+    } else {
+      play1[1] <- which(s1 == init1)
+    }
+    if (is.null(init2)) {
+      play2[1] <- sample(1:n2, size = 1)
+    } else {
+      play2[1] <- which(s2 == init2)
+    }
+
+    for (t in 2:np2) {
+      if (t %% 2 != 0) {
         ## Player 1
         if (stats::runif(1) < omega) {
-          play1[i] <- play1[i - 1]
+          play1[t] <- play1[t - 1]
         } else {
-          df1 <- game$df %>%
-            dplyr::filter(s2 == play2[i - 1]) %>%
-            dplyr::filter(payoff1 == max(payoff1))
-          if (nrow(df1) > 1) df1 <- dplyr::slice_sample(df1, 1)
-          play1[i] <- df1$s1[1]
+          pi1t <- pi1[, play2[t -1]]
+          pi1t <- pi1t == max(pi1t)
+          if (sum(pi1t) == 1) {
+            play1[t] <- which(pi1t)
+          } else {
+            play1[t] <- sample(which(pi1t), size = 1)
+          }
         }
         ## Player 2
-        play2[i] <- play2[i - 1]
+        play2[t] <- play2[t - 1]
       } else {
         ## Player 1
-        play1[i] <- play1[i - 1]
+        play1[t] <- play1[t - 1]
         ## Player 2
         if (stats::runif(1) < omega) {
-          play2[i] <- play2[i - 1]
+          play2[t] <- play2[t - 1]
         } else {
-          df2 <- game$df %>%
-            dplyr::filter(s1 == play1[i - 1]) %>%
-            dplyr::filter(payoff2 == max(payoff2))
-          if (nrow(df2) > 1) df2 <- dplyr::slice_sample(df2, 1)
-          play2[i] <- df2$s2[1]
+          pi2t <- pi2[play1[t -1], ]
+          pi2t <- pi2t == max(pi2t)
+          if (sum(pi2t) == 1) {
+            play2[t] <- which(pi2t)
+          } else {
+            play2[t] <- sample(which(pi2t), size = 1)
+          }
         }
       }
     }
+    play1 <- s1[play1]
+    play2 <- s2[play2]
     play2[1] <- NA
 
   } else if (game$type == "char_function") {
@@ -90,7 +108,7 @@ sim_game_abr <- function(game,
     if (is.null(init2)) play2[1] <- stats::runif(1, min = s2[1], max = s2[2])
     else play2[1] <- init2
 
-    for (i in 2:n2) {
+    for (i in 2:np2) {
       if (i %% 2 != 0) {
         ## Player 1
         if (stats::runif(1) < omega) {
@@ -146,7 +164,7 @@ sim_game_abr <- function(game,
     if (is.null(init2)) play2[1] <- stats::runif(1, min = s2[1], max = s2[2])
     else play2[1] <- init2
 
-    for (i in 2:n2) {
+    for (i in 2:np2) {
       if (i %% 2 != 0) {
         ## Player 1
         if (stats::runif(1) < omega) {
@@ -208,8 +226,8 @@ sim_game_abr <- function(game,
     play2[1] <- NA
   }
 
-  return(data.frame(play1  = play1,
-                    play2  = play2,
-                    period = 1:n2,
+  return(data.frame(play1 = play1,
+                    play2 = play2,
+                    period = 1:np2,
                     moved = rep(game$player, n_periods)))
 }