+++ R/sim_game_sbr.R
@@ -47,38 +47,48 @@ sim_game_sbr <- function(game,
     s1 <- game$strategy$s1
     s2 <- game$strategy$s2
 
-    # for the first round
-    if (is.null(init1)) play1[1] <- sample(s1, size = 1)
-    else play1[1] <- init1
-    if (is.null(init2)) play2[1] <- sample(s2, size = 1)
-    else play2[1] <- init2
+    n1 <- length(s1)
+    n2 <- length(s2)
 
-    for (i in 2:n_periods) {
+    pi1 <- game$mat$matrix1
+    pi2 <- game$mat$matrix2
+
+    # for the first round
+    if (is.null(init1)) {
+      play1[1] <- sample(1:n1, size = 1)
+    } else {
+      play1[1] <- which(s1 == init1)
+    }
+    if (is.null(init2)) {
+      play2[1] <- sample(1:n2, size = 1)
+    } else {
+      play2[1] <- which(s2 == init2)
+    }
+
+    for (t in 2:n_periods) {
       ## Player 1
       if (stats::runif(1) < omega) {
-        play1[i] <- play1[i - 1]
+        play1[t] <- play1[t - 1]
       } else {
-        p <- game$df %>%
-          dplyr::filter(s2 == play2[i - 1]) %>%
-          dplyr::pull(payoff1)
+        p <- pi1[, play2[t - 1]]
         p <- exp(lambda * p)
         p <- p / sum(p)
-        play1[i] <- sample(s1, size = 1, prob = p)
+        play1[t] <- sample(1:n1, size = 1, prob = p)
       }
 
       ## Player 2
       if (stats::runif(1) < omega) {
-        play2[i] <- play2[i - 1]
+        play2[t] <- play2[t - 1]
       } else {
-        ###
-        p <- game$df %>%
-          dplyr::filter(s1 == play1[i - 1]) %>%
-          dplyr::pull(payoff2)
+        p <- pi2[play2[t - 1], ]
         p <- exp(lambda * p)
         p <- p / sum(p)
-        play2[i] <- sample(s2, size = 1, prob = p)
+        play2[t] <- sample(1:n2, size = 1, prob = p)
       }
     }
+
+    play1 <- s1[play1]
+    play2 <- s2[play2]
 
   } else if (game$type == "char_function") {